Incorporates Marek's feedback
